+++ Modules/Intent.Modules.Application.MediatR/Templates/QueryModels/QueryModelsTemplatePartial.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Intent.Engine;
+using Intent.Modelers.Services.Api;
 using Intent.Modelers.Services.CQRS.Api;
 using Intent.Modules.Application.DependencyInjection.MediatR;
 using Intent.Modules.Application.MediatR.Settings;
@@ -58,12 +59,34 @@ namespace Intent.Modules.Application.MediatR.Templates.QueryModels
                     // get the last property which has no value. All items occuring before this cannot have a default value set in the constructor
                     var lastNonNullable = Model.Properties.LastOrDefault(p => string.IsNullOrEmpty(p.Value))?.InternalElement.Order ?? 0;
 
+                    List<string> nulledFields = [];
                     foreach (var property in Model.Properties)
                     {
-                        ctor.AddParameter(GetTypeName(property), property.Name.ToParameterName(), param =>
+                        // should the default value be set, based on the position of it as a argument
+                        var setDefaultValue = property.ShouldSetDefaultValue(lastNonNullable);
+                        // set the type
+                        var typeValue = property.GetTypeReferenceName(setDefaultValue, this);
+
+                        ctor.AddParameter(typeValue, property.Name.ToParameterName(), param =>
                         {
                             param.AddMetadata("model", property);
-                            param.IntroduceProperty(prop =>
+
+                            // only parameters with a value AFTER the last parameter with a value get the value specified
+                            if (property.InternalElement.Order >= lastNonNullable && !string.IsNullOrEmpty(property.Value))
+                            {
+                                param.WithDefaultValue(property.Value);
+
+                                // if is a collection, with a default value, set to null instead
+                                if (property.TypeReference?.IsCollection ?? false)
+                                {
+                                    param.WithDefaultValue("null");
+                                    nulledFields.Add(property.Id);
+                                }
+                            }
+
+                            // AddProperty is used instead of IntroduceProperty as the property and the parameter might not have the same type
+                            // One could be non-nullable and the other not, specifically when its a collection with default value
+                            @class.AddProperty(GetTypeName(property), property.Name.ToPropertyName(), prop =>
                             {
                                 prop.TryAddXmlDocComments(property.InternalElement);
                                 prop.RepresentsModel(property);
@@ -73,13 +96,15 @@ namespace Intent.Modules.Application.MediatR.Templates.QueryModels
                                 {
                                     prop.WithComments(xmlComments: $"/// <example>{property.GetStereotype("OpenAPI Settings").GetProperty("Example Value")?.Value}</example>");
                                 }
+
+                                // Do the assigmentment in the constructor, if the parameter has a default value, we need to use the null-coalescing operator to assign the default value to the property if the parameter is null
+                                var rhs = setDefaultValue && nulledFields.Contains(property.Id) ? $"{property.Name.ToParameterName()} ?? {property.Value}" :
+                                    property.Name.ToParameterName();
+                                var assignmentStatement = new CSharpFieldAssignmentStatement(prop.Name, rhs);
+                                ctor.AddStatement(assignmentStatement);
                             });
 
-                            // only parameters with a value AFTER the last parameter with a value get the value specified
-                            if (property.InternalElement.Order >= lastNonNullable && !string.IsNullOrEmpty(property.Value))
-                            {
-                                param.WithDefaultValue(property.Value);
-                            }
+                            
                         });